fix(seed-questions): render seed questions and shuffle
